+++ src/features/characters/CharacterSlot.tsx
