+++ src/features/expedition/components/CharacterPanel.tsx
